Moves business logic of HomePage to useHomePageViewModel
